refactor(presenter): organized presenter code with more files
